// src/app/pages/b/credit-info/Models/earlier-loan-info.ts
export class EarlierLoanInfo {
    ID: number;
    CLIENT_ID: number;
    CREDIT_INFORMATION_ID: number;
    BANK_NAME: string;
    BRANCH_NAME: string;
    LOAN_AMOUNT: number;
    PURPOSE_OF_LOAN: string;
    LOAN_TAKEN_DATE: any;
    ACCOUNT_CLOSE_DATE: any;
    IS_SUB: boolean;
    DUEDATE: any;
    REPAYMENT_DATE: any;
    REMARK: string; 
    YEAR:number 
    LOAN_TYPE_NAME_KN=''
}